Update theme for easy development
- Update sphinx so it's up to the newest version
- Update the hole theme to use the npm webpack package for development of the theme.
The webpack package is what the latest sphinx themes uses.
Running the dev server works fine read in the readme file how to set up the dev server.
Testing the theme changes should be done in first hand on the local dev server but if you ant to test it with the shuup guide there is documentation for that to.
- Change nav bar color to be black with white text